refactor getcollectionsprovider
